requirements.txt
Flask==2.2.3
Flask_Bcrypt==1.0.1
Flask_Login==0.6.2
Flask_RESTful==0.3.9
flask_sqlalchemy==3.0.3
Flask_WTF==1.1.1
jsonify==0.5
PyJWT==2.6.0
PyJWT==2.6.0
Requests==2.28.2
WTForms==3.0.1
flask_session
psycopg2
email_validator
